pureber: rethink decoding strategy
instead of passing a context object to each fromBER function,
the top level object needs to decide how its attributes shall
be decoded.

This feels more natural, since the object class knows the context in
which its attributes shall be decoded.

Additionaly, this allowes more natural code paths which are easier to type.